Make: Consolidated Updates to C# Bindings
This commit introduce changes in C# binding.
It contains fixes, massive code refactor, build scripts, docs.

In particular these changes:

- Make: .yml workflow to create artifacts usearch libs
- Make: .yml workflows for build, test, pack, send to nuget
- Make: Build scripts
- Refactor: Change names of dirs, project, code
- Docs: README files with quickstart and for nuget
