+++ .github/workflows/csharp-build-and-test.yml
@@ -1,49 +1,71 @@
-# This workflow builds USearch native libraries, then build and test C# USearch wrapper
-# Currently for development purposes an if condition specified below to work only fork 
+# This workflow builds the USearch native libraries, then builds and tests the C# USearch wrapper.
+# Currently, for development purposes, an 'if' condition is specified below to work only on forks.
+# TODO: replace branches ref in prod
+# TODO: remove `if` conditions in jobs in prod
+name: Test C#
 
-name: Test C# wrapper
-
+# on: workflow_dispatch
 on:
   push:
     branches: ["160-feature-c-bindings-and-semantic-kernel-integration"]
 
 env:
-  LINUX_OSX_SCRIPT: build_and_test.sh
-  WINDOWS_SCRIPT: build_and_test.cmd
+  CHECKOUT_BRANCH: 160-feature-c-bindings-and-semantic-kernel-integration
 
 jobs:
   matrix-test:
-      if: github.repository == 'AleksandrKent/usearch'
-      runs-on: ${{ matrix.os }}  
-      strategy:
-        fail-fast: false
-        matrix:
-          os: [ubuntu-latest, windows-latest, macos-latest]
-      steps:
-        - uses: actions/checkout@v3
-          with:
-            ref: "160-feature-c-bindings-and-semantic-kernel-integration"
-        - run: |
-            git submodule update --init --recursive
-          shell: bash
+    if: github.repository == 'AleksandrKent/usearch'
+    runs-on: ${{ matrix.os }}
+    strategy:
+      fail-fast: false
+      matrix:
+        os: [ubuntu-latest, windows-latest, macos-latest]
+    steps:
+      - uses: actions/checkout@v3
+        with:
+          ref: ${{ env.CHECKOUT_BRANCH }}
+      - run: |
+          git submodule update --init --recursive
+        shell: bash
 
-        - name: Setup .NET
-          uses: actions/setup-dotnet@v3
-          with:
-            dotnet-version: 6.0.x
+      - name: Setup .NET
+        uses: actions/setup-dotnet@v3
+        with:
+          dotnet-version: 6.0.x
 
-        # Ubuntu and macOS steps
-        - name: Run ${{ env.LINUX_OSX_SCRIPT }} on Ubuntu or macOS
-          if: matrix.os == 'ubuntu-latest' || matrix.os == 'macos-latest'
-          run: |
-            cd ./csharp
-            "./${{ env.LINUX_OSX_SCRIPT }}"
-          shell: bash
-        
-        # Windows step
-        - name: Run ${{ env.WINDOWS_SCRIPT }} on Windows
-          if: matrix.os == 'windows-latest'
-          run: |
-            cd .\csharp
-            ${{ env.WINDOWS_SCRIPT }}
-          shell: cmd
+      # Ubuntu and macOS steps
+      - name: Build native lib on Ubuntu or macOS
+        if: matrix.os == 'ubuntu-latest' || matrix.os == 'macos-latest'
+        run: |
+          ./build_lib.sh
+        shell: bash
+        working-directory: ${{ github.workspace }}/csharp/eng/
+
+      # Windows steps
+
+      # Use cl.exe https://www.appveyor.com/docs/lang/cpp/#visual-studio-2017
+      - name: Custom MSVC build
+        if: matrix.os == 'windows-latest'
+        run: |
+          call "C:\Program Files\Microsoft Visual Studio\2022\Enterprise\VC\Auxiliary\Build\vcvars64.bat"
+          where cl
+          cl /EHsc /std:c++14 /O2 /LD /Fe:libusearch_c.dll lib.cpp /I. /I../include/ /I../fp16/include/ /I../robin-map/include/
+        working-directory: ${{ github.workspace }}/c
+        shell: cmd
+
+      - name: Copy dll on Windows to lib dir
+        if: matrix.os == 'windows-latest'
+        run: |
+          mkdir -p "${GITHUB_WORKSPACE}/csharp/lib/runtimes/win-x64/native"
+          cp "${GITHUB_WORKSPACE}/c/libusearch_c.dll" "${GITHUB_WORKSPACE}/csharp/lib/runtimes/win-x64/native"
+        shell: bash
+
+      - name: Dotnet build
+        run: dotnet build -c Release
+        shell: bash
+        working-directory: ${{ github.workspace }}/csharp/
+
+      - name: Dotnet test
+        run: dotnet test -c Release --no-build
+        shell: bash
+        working-directory: ${{ github.workspace }}/csharp/
