+++ .github/workflows/update_version.sh
@@ -10,4 +10,5 @@ echo $1 > VERSION &&
     sed -i "s/^\(#define USEARCH_VERSION_PATCH \).*/\1$(echo "$1" | cut -d. -f3)/" ./include/usearch/index.hpp &&
     sed -i "s/<version>[0-9]\+\.[0-9]\+\.[0-9]\+/<version>$1/" README.md &&
     sed -i "s/version = {[0-9]\+\.[0-9]\+\.[0-9]\+}/version = {$1}/" README.md &&
-    sed -i "s/version=\".*\"/version=\"$1\"/" wasmer.toml
+    sed -i "s/version=\".*\"/version=\"$1\"/" wasmer.toml &&
+    sed -i "s/>[0-9]*\.[0-9]*\.[0-9]*<\/Version>/>$1<\/Version>/" ./csharp/Directory.Build.props
