+++ csharp/nuget/NUGET.md
@@ -1,0 +1,11 @@
+**USearch.NET** is a wrapper for the USearch vector search engine, offering fast searches, multiple metrics, and features like SIMD-optimized metrics.
+
+**Supported metrics**: Euclidean, Angular, Jaccard, Hamming, Haversine, User-Defined Metrics
+
+**Platforms**: Linux, MacOS, Windows
+
+# Getting Started ⚡
+
+- [Quickstart](https://github.com/unum-cloud/usearch/blob/main/csharp/README.md)
+- Join the [Discord community](https://discord.gg/A6wxt6dS9j)
+- Check out the [GitHub repository](https://github.com/unum-cloud/usearch)
