--- csharp/src/LibUSearch/LibUSearch.csproj
@@ -1,16 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-    <PropertyGroup>
-        <TargetFramework>netstandard2.0</TargetFramework>
-        <AssemblyName>Cloud.Unum.USearch</AssemblyName>
-        <RootNamespace>Cloud.Unum.USearch</RootNamespace>
-        <LangVersion>10</LangVersion>
-    </PropertyGroup>
-
-    <!-- Explicitly clear output folders to avoid issues with out-of-date files like dlls-->
-    <Target Name="PostClean" AfterTargets="Clean">
-        <RemoveDir Directories="$(BaseIntermediateOutputPath)" /><!-- obj -->
-        <RemoveDir Directories="$(BaseOutputPath)" /><!-- bin -->
-    </Target>
-
-</Project>
