+++ csharp/src/USearchNET.Tests/Usings.cs
@@ -1,0 +1,1 @@
+global using Xunit;
